[vim] new, experimental trailing space function and menu item
Visibility of listchars is handled by g:solarized_visibility
but on the cursorline they can still be hard to spot if cursorline
is on due to the cursorline background color.

This new trailing spaces function, available in the menu or via
emenu, should change the trailing space highlighting on only
the cursorline. It doesn't work in call cases as, for example,
the comments pattern in most syntaxes will override it.