Link to the most popular implementation for Emacs
Several solarized implementations (and a fork) exist for Emacs:
http://www.melpa.org/#/?q=solar&sort=downloads&asc=false

The implementation that this repository used to link to is the least
popular among those.  Link to the most popular one instead.  I am a
bit surprised that it is not at least a magnitude more popular than
the second most popular one, but it is much more actively maintained
(68 vs 6 contributors, 601 vs 194 commits).

This also removes the bundled copy of the the older implementation
without bundling a copy of the newer implementation.  It seems those
bundled copies are not getting updated and I don't volunteer to do
it, so it seems better to just link to the up-to-date standalone
repository.

I have contacted the maintainers of the other implementations, last
here: https://github.com/melpa/melpa/pull/5780 (which also provides
some background information in case you are interested).  @sellout,
the author of the implementation previously linked to, never replied.